+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Messages with general context from the Checkout API are now displayed in toasts.
+
 ## [0.9.0] - 2019-09-23
 
 ## Changed